+++ frontend/package.json
@@ -18,6 +18,7 @@
     "node-sass": "^4.14.1",
     "vue": "^3.1.0",
     "vue-window-size": "^1.1.2",
+    "vuedraggable": "^4.1.0",
     "vueform": "^2.0.0",
     "vuex": "^4.0.2"
   },
