+++ frontend/src/components/Toolbar.vue
@@ -21,15 +21,40 @@
         <strong>Add layer</strong>
       </el-button>
 
-      <el-select v-model="currentLayer">
-        <el-option
-          v-for="(layer, index) in layers"
-          :key="index"
-          :label="layer"
-          :value="layer"
-        >
-        </el-option>
-      </el-select>
+      <el-dropdown split-button :hide-on-click="false" v-if="layers != null">
+        {{ currentLayer }}
+        <template #dropdown>
+          <el-dropdown-menu>
+            <draggable
+              v-model="layers"
+              tag="transition-group"
+              handle=".handle"
+              :item-key="(element) => element"
+            >
+              <template #item="{element}">
+                <div
+                  class="layer-item"
+                  :class="{
+                    selected: currentLayer == element,
+                    first: layers[0] == element,
+                    last: layers[layers.length - 1] == element,
+                  }"
+                >
+                  <img
+                    class="handle"
+                    :src="require('../assets/tools/d-caret.svg')"
+                  />
+                  <span
+                    style="cursor: pointer"
+                    @click="currentLayer = element"
+                    >{{ element }}</span
+                  >
+                </div>
+              </template>
+            </draggable>
+          </el-dropdown-menu>
+        </template>
+      </el-dropdown>
     </div>
     <div style="display: flex; align-items: center;">
       <el-tag v-if="!isOnline" class="connBadge" type="danger"
@@ -54,17 +79,36 @@
 import { key, State } from "@/store";
 import { computed, defineComponent } from "vue";
 import { useStore } from "vuex";
+import draggable from "vuedraggable";
 import "element-plus/dist/index.css";
+
+interface layerData {
+  id: string;
+  zIndex: number;
+}
 
 export default defineComponent({
   name: "Toolbar",
-  emits: ["addLayer"],
+  emits: ["addLayer", "layersReordered"],
+  components: {
+    draggable,
+  },
   setup(props, { emit }) {
     console.log("Toolbar setup");
     const store = useStore<State>(key);
 
-    const layers = computed(() => {
-      return store.state.layers.map((layer) => layer.attrs.id);
+    const layers = computed({
+      get: function() {
+        const tempLayers: layerData[] = [];
+        for (const layer of store.state.layers)
+          tempLayers.push({ id: layer.id(), zIndex: layer.zIndex() });
+        return tempLayers
+          .sort((layer1, layer2) => layer2.zIndex - layer1.zIndex)
+          .map((layer) => layer.id);
+      },
+      set: function(layers: string[]) {
+        emit("layersReordered", layers);
+      },
     });
 
     const currentLayer = computed({
@@ -103,12 +147,15 @@ export default defineComponent({
       emit("addLayer");
     };
 
+    let drag = false;
+
     return {
       layers,
       currentLayer,
       isOnline,
       currentTool,
       addLayer,
+      drag,
     };
   },
   mounted() {
@@ -160,5 +207,36 @@ export default defineComponent({
     font-size: 12px;
     width: 80px;
   }
+}
+.el-dropdown-menu {
+  background-color: #f6f6f6;
+}
+.layer-item {
+  margin: 0px 10px;
+  padding: 12px 10px;
+  width: 80px;
+  border: 1px lightgray solid;
+  background-color: #fff;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  font-size: 14px;
+}
+.handle {
+  float: left;
+  width: 16px;
+  height: 16px;
+  margin-right: 3px;
+}
+.selected {
+  border: 1px blue solid;
+}
+.first {
+  border-top-left-radius: 3px;
+  border-top-right-radius: 3px;
+}
+.last {
+  border-bottom-left-radius: 3px;
+  border-bottom-right-radius: 3px;
 }
 </style>
