+++ frontend/src/store.ts
@@ -49,20 +49,18 @@ export const store = createStore<State>({
     },
 
     setCurrentLayer(state, layer) {
-      console.log(layer);
       state.currentLayer = layer;
-    },
-    setLayers(state, layers) {
-      state.layers = layers;
     },
     setStage(state, stage) {
       state.stage = stage;
+    },
+    setLayers(state, layers) {
+      state.layers = layers;
     },
     addLayer(state, layer) {
       if (state.stage !== undefined) {
         state.stage.add(layer);
         state.layers = [...state.layers, layer];
-        console.log(state.layers);
       }
     },
     setCurrentTool(state, tool) {
